pyproject.toml
[tool.poetry]
name = "imager-bot"
version = "0.1.0"
description = ""
authors = ["Mikhail Drozdov <deathlokmike@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
selenium = "^4.20.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
validators = "^0.28.1"
loguru = "^0.7.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
